requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.7
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Jinja2==3.1.2 